refactor(guest): move ResolvedSecurityPolicy out of capabilities.rs
capabilities.rs's own module doc scoped it to "Linux capabilities for
container processes", but ResolvedSecurityPolicy bundles three fields —
capabilities is only one of them, readonly_paths/sys_mount_options are
mount paths and options, unrelated to capability resolution. Two of its
three consumers (start.rs, lifecycle.rs) only ever wanted the bundle, not
CapabilitySet itself, and vice versa for command.rs/zygote.rs — the file
was doing double duty for two different call patterns.

Before: capabilities.rs { CAPABILITIES_BY_NUMBER, CapabilitySet,
  ResolvedSecurityPolicy (capabilities + readonly_paths + sys_mount_options) }
After: capabilities.rs { CAPABILITIES_BY_NUMBER, CapabilitySet } — back to
  matching its own doc comment
  security_policy.rs (new) { ResolvedSecurityPolicy, using
  capabilities::CapabilitySet as one of its three fields }

Pure move: no logic changes. CapabilitySet's test-only len()/contains()
went from private to pub(crate) so security_policy.rs's tests (moved
along with the struct) can still reach them from a sibling module.
