+++ src/guest/src/container/mod.rs
@@ -68,6 +68,8 @@ mod kill;
 #[cfg(target_os = "linux")]
 mod lifecycle;
 #[cfg(target_os = "linux")]
+mod security_policy;
+#[cfg(target_os = "linux")]
 mod spec;
 #[cfg(target_os = "linux")]
 mod start;
@@ -79,10 +81,10 @@ pub(crate) mod user_profile;
 pub(crate) mod zygote;
 
 #[cfg(target_os = "linux")]
-pub(crate) use capabilities::ResolvedSecurityPolicy;
-#[cfg(target_os = "linux")]
 pub(crate) use command::SpawnResult;
 #[cfg(target_os = "linux")]
 pub use lifecycle::Container;
+#[cfg(target_os = "linux")]
+pub(crate) use security_policy::ResolvedSecurityPolicy;
 #[cfg(target_os = "linux")]
 pub use spec::{ContainerDevices, UserMount};
