+++ src/guest/src/container/spec.rs
@@ -2,7 +2,8 @@
 //!
 //! Creates OCI-compliant runtime specifications following the runtime-spec standard.
 
-use super::capabilities::{CapabilitySet, ResolvedSecurityPolicy};
+use super::capabilities::CapabilitySet;
+use super::security_policy::ResolvedSecurityPolicy;
 use boxlite_shared::errors::{BoxliteError, BoxliteResult};
 use boxlite_shared::ContainerDevice as ProtoContainerDevice;
 use std::os::unix::fs::{FileTypeExt, MetadataExt};
